+++ scdtr/include/controller.hpp
@@ -35,7 +35,7 @@ class Controller {
      *
      * @return The PWM signal.
      */
-    uint8_t compute_pwm_signal(float luminusity, uint32_t current_time);
+    uint16_t compute_pwm_signal(float luminusity, uint32_t current_time);
 
    private:
     bool anti_windup;
